ShowTools: extract add+update and get tools.
