+++ app/src/main/java/com/battlelancer/seriesguide/sync/ShowSync.kt
@@ -6,14 +6,13 @@ import android.text.format.DateUtils
 import com.battlelancer.seriesguide.provider.SgRoomDatabase
 import com.battlelancer.seriesguide.sync.SgSyncAdapter.UpdateResult
 import com.battlelancer.seriesguide.sync.SyncOptions.SyncType
-import com.battlelancer.seriesguide.ui.shows.ShowTools
-import com.battlelancer.seriesguide.ui.shows.ShowTools2
-import com.battlelancer.seriesguide.ui.shows.ShowTools2.UpdateResult.ApiErrorRetry
-import com.battlelancer.seriesguide.ui.shows.ShowTools2.UpdateResult.ApiErrorStop
-import com.battlelancer.seriesguide.ui.shows.ShowTools2.UpdateResult.DatabaseError
-import com.battlelancer.seriesguide.ui.shows.ShowTools2.UpdateResult.DoesNotExist
-import com.battlelancer.seriesguide.ui.shows.ShowTools2.UpdateResult.Success
 import com.battlelancer.seriesguide.util.TimeTools
+import com.battlelancer.seriesguide.util.shows.AddUpdateShowTools
+import com.battlelancer.seriesguide.util.shows.AddUpdateShowTools.UpdateResult.ApiErrorRetry
+import com.battlelancer.seriesguide.util.shows.AddUpdateShowTools.UpdateResult.ApiErrorStop
+import com.battlelancer.seriesguide.util.shows.AddUpdateShowTools.UpdateResult.DatabaseError
+import com.battlelancer.seriesguide.util.shows.AddUpdateShowTools.UpdateResult.DoesNotExist
+import com.battlelancer.seriesguide.util.shows.AddUpdateShowTools.UpdateResult.Success
 import com.uwetrottmann.androidutils.AndroidUtils
 import timber.log.Timber
 import kotlin.math.pow
@@ -44,7 +43,6 @@ class ShowSync(
     @SuppressLint("TimberExceptionLogging")
     fun sync(
         context: Context,
-        showTools: ShowTools,
         currentTime: Long,
         progress: SyncProgress
     ): UpdateResult? {
@@ -53,10 +51,11 @@ class ShowSync(
         val showsToUpdate = getShowsToUpdate(context, currentTime) ?: return null
         Timber.d("Updating %d show(s)...", showsToUpdate.size)
 
+        val showTools = AddUpdateShowTools(context)
         var networkErrors = 0
         for (showId in showsToUpdate) {
             // Try to update this show.
-            var result: ShowTools2.UpdateResult
+            var result: AddUpdateShowTools.UpdateResult
             do {
                 // Shortcut to stop updating if connectivity is lost.
                 if (!AndroidUtils.isNetworkConnected(context)) {
