+++ app/src/main/java/com/battlelancer/seriesguide/traktapi/TraktTools2.kt
@@ -2,9 +2,9 @@ package com.battlelancer.seriesguide.traktapi
 
 import android.content.Context
 import com.battlelancer.seriesguide.SgApp
-import com.battlelancer.seriesguide.ui.shows.ShowTools2.ShowResult
 import com.battlelancer.seriesguide.util.Errors
 import com.battlelancer.seriesguide.util.isRetryError
+import com.battlelancer.seriesguide.util.shows.AddUpdateShowTools.ShowResult
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.Ok
 import com.github.michaelbull.result.Result
